- Change time format - Change the success message to automatically proceed to the next page

+++ resources/views/pages/user-manual-timein.blade.php
@@ -49,6 +49,8 @@
                             icon: 'success',
                             title: 'Success',
                             text: 'Attendance registered successfully',
+                            timer: 2000, // 2 seconds
+                            showConfirmButton: false
                         }).then(() => {
                             sessionStorage.setItem('name', response.name);
                             sessionStorage.setItem('id_number', response.id_number);
